Move getColor to own file for #14
